+++ lib/src/generator/data_layer_generator.dart
@@ -45,27 +45,27 @@ class DataLayerGenerator {
     for (final method in config.methods) {
       switch (method) {
         case 'get':
-          methods.add('  Future<$entityName> get(${config.idType} id);');
+          methods.add('  Future<$entityName> get(${config.queryFieldType} ${config.queryField});');
           break;
         case 'getList':
-          methods.add('  Future<List<$entityName>> getList();');
+          methods.add('  Future<List<$entityName>> getList(ListQueryParams params);');
           break;
         case 'create':
           methods
               .add('  Future<$entityName> create($entityName $entityCamel);');
           break;
         case 'update':
-          methods
-              .add('  Future<$entityName> update($entityName $entityCamel);');
+          methods.add(
+              '  Future<${config.name}> update(UpdateParams<${config.useMorphy ? "${config.name}Patch" : "Partial<${config.name}>"}> params);');
           break;
         case 'delete':
-          methods.add('  Future<void> delete(${config.idType} id);');
+          methods.add('  Future<void> delete(DeleteParams<$entityName> params);');
           break;
         case 'watch':
-          methods.add('  Stream<$entityName> watch(${config.idType}? id);');
+          methods.add('  Stream<$entityName> watch(${config.queryFieldType}? ${config.queryField});');
           break;
         case 'watchList':
-          methods.add('  Stream<List<$entityName>> watchList();');
+          methods.add('  Stream<List<$entityName>> watchList(ListQueryParams params);');
           break;
       }
     }
@@ -137,15 +137,15 @@ ${methods.join('\n')}
         case 'get':
           methods.add('''
   @override
-  Future<$entityName> get(${config.idType} id) {
-    return _dataSource.get(id);
+  Future<$entityName> get(${config.queryFieldType} ${config.queryField}) {
+    return _dataSource.get(${config.queryField});
   }''');
           break;
         case 'getList':
           methods.add('''
   @override
-  Future<List<$entityName>> getList() {
-    return _dataSource.getList();
+  Future<List<$entityName>> getList(ListQueryParams params) {
+    return _dataSource.getList(params);
   }''');
           break;
         case 'create':
@@ -158,29 +158,30 @@ ${methods.join('\n')}
         case 'update':
           methods.add('''
   @override
-  Future<$entityName> update($entityName $entityCamel) {
-    return _dataSource.update($entityCamel);
-  }''');
+  Future<${config.name}> update(UpdateParams<${config.useMorphy ? "${config.name}Patch" : "Partial<${config.name}>"}> params) {
+    return _dataSource.update(params);
+  }
+''');
           break;
         case 'delete':
           methods.add('''
   @override
-  Future<void> delete(${config.idType} id) {
-    return _dataSource.delete(id);
+  Future<void> delete(DeleteParams<$entityName> params) {
+    return _dataSource.delete(params);
   }''');
           break;
         case 'watch':
           methods.add('''
   @override
-  Stream<$entityName> watch(${config.idType}? id) {
-    return _dataSource.watch(id);
+  Stream<$entityName> watch(${config.queryFieldType}? ${config.queryField}) {
+    return _dataSource.watch(${config.queryField});
   }''');
           break;
         case 'watchList':
           methods.add('''
   @override
-  Stream<List<$entityName>> watchList() {
-    return _dataSource.watchList();
+  Stream<List<$entityName>> watchList(ListQueryParams params) {
+    return _dataSource.watchList(params);
   }''');
           break;
       }