+++ test_project/lib/src/domain/entities/multi_field_entity.dart
@@ -1,0 +1,4 @@
+class MultiFieldEntity {
+  final String id;
+  final int x, y, z;
+}